feat: Complete tasks TASK-001 through TASK-028
- TASK-001: DB production hardening (busy timeout, version_number, FK test)
- TASK-002: Extract withTx helper, replace 14 inline transaction blocks
- TASK-003: fmt.Printf → structured logging, scanAll pattern, pending.go int64 fix
- TASK-004: Consolidate duplicate test helpers (SetupTestDB, remove setupBenchmarkDB)
- TASK-005: Fix *sql.DB type-assertion anti-pattern in stores
- TASK-006: Add validation tags and input validation to models/stores
- TASK-007: Add consistent error sentinels (ErrGroupNotFound, ErrPolicyNotFound, etc.)
- TASK-008: Extract execUpdate/softDelete/getNameByID helpers, DRY cleanup
- TASK-009: JWT key rotation, iss/aud validation, PBKDF2 600k, gcmTagSize const
- TASK-010: Fix flaky TestTokenExpiration, consolidate middleware tests, add tests
- TASK-011: Log cleanup worker panic recovery, sync.Mutex, remove redundant test
- TASK-012: Remove init() metrics auto-registration, add RegisterMetrics()
- TASK-013: Normalize time constants, consolidate OfflineThreshold, add docs
- TASK-014: Integrate updater goroutine into errgroup lifecycle
- TASK-015: Fix flush window, consolidate iptables backup, cache hasDocker()
- TASK-016: SSE error propagation through errgroup, Run() decomposition
- TASK-017: Agent fixes (sort. SliceStable, context, disableService, writeTempFile)
- TASK-018: Fix SQL injection (parameterized queries), extract DB view, fix double defer
- TASK-019: Decompose generateIptablesPayload into 8 sub-functions, unify writeRules
- TASK-020: Fix self-referencing filter (CIDR), invertPortMatch, transitive RecompileAffectedPeers
- TASK-021: CIDR normalization, ipset name validation, HKDF signing, version_number replay protection
- TASK-022: Extract moduleHandlers helper, log unrecognized tokens, add completeness validation
- TASK-023: Fix benchmark DB cleanup, rune overflow, double defer cleanup(), test fixes
- TASK-024: Move QueuePeerChange inside transaction, extract sqlutil, fix row.Close logging
- TASK-025: Create internal/resolve package, unify broadcast resolution, deduplicate special targets
- TASK-026: Delete unused templates, remove SanitizeAlertInputStrict and no-op evaluators
- TASK-027: Add sync.RWMutex to spike detector, fix per-peer coverage gap
- TASK-028: SMTP retry with backoff, configurable HELO, DRY HTML generation

+++ internal/db/pending.go
@@ -87,12 +87,12 @@ func DeletePendingBundlePreview(ctx context.Context, database Querier, peerID in
 	return err
 }
 
-func DeletePendingChangeForEntity(ctx context.Context, database Querier, peerID int64, changeType string, changeID int) error {
+func DeletePendingChangeForEntity(ctx context.Context, database Querier, peerID int, changeType string, changeID int) error {
 	_, err := database.ExecContext(ctx, "DELETE FROM pending_changes WHERE peer_id = ? AND change_type = ? AND change_id = ?", peerID, changeType, changeID)
 	return err
 }
 
-func CountPendingChangesForPeer(ctx context.Context, database Querier, peerID int64) (int, error) {
+func CountPendingChangesForPeer(ctx context.Context, database Querier, peerID int) (int, error) {
 	var count int
 	err := database.QueryRowContext(ctx, "SELECT COUNT(*) FROM pending_changes WHERE peer_id = ?", peerID).Scan(&count)
 	if err != nil {
